Implemented pathfinding. Weighting should be implemented for ground, however have not implemented impassable objects (aka other enemies blocking your path). You can't step on them, but you essentially "pass through" them at this point. I will look into this again later. Please test this, as there were a LOT of changes made to get it working. Also an FYI, rebuild after Karls update, as it did not change for me until after a full rebuild.

+++ src/GameEngine/Character.cpp
@@ -11,6 +11,7 @@
  * Mike Malyuk, February 14 2007 | Changed attacking to cout important info. Changed Defence calc by dividing by 2
  *                               | Protected against dealing "negative damage" or healing the defender.
  * Mike Malyuk, February 14 2007 | Levelling up occurs at 100 exp now instead of greater than 100
+ * Mike Malyuk, March 10 2007    | Added explicit cast so no warning, the int cutting is intentional
  */
 
 #include <util.h>
@@ -134,7 +135,7 @@ void Character::Attack(Character* another)
     }
     if(killed == true)
     {
-        mExp = mExp + ((another->GetLevel()*1.0)/mLevel)*100;
+        mExp = (int)(mExp + ((another->GetLevel()*1.0)/mLevel)*100);
         if(mExp >= 100)
         {
             LevelUp();
